Code accounts for pv exception
