+++ constants.py
@@ -1,5 +1,7 @@
 from PyQt5.QtGui import QColor
 from scLinac import LINACS, Linac, Cavity
+from fault import faults
+from epics import PV
 
 greenFillColor = QColor(201,255,203)
 neonGreenBorderColor = QColor(46,248,10)
@@ -32,8 +34,30 @@ shapeParameterDict = {0: ShapeParameters(greenFillColor, neonGreenBorderColor,
 STATUS_SUFFIX = "CUDSTATUS"
 SEVERITY_SUFFIX = "CUDSEVR"
 
+class DisplayCavity(Cavity, object):
+    def __init__(self, cavityNum, rackObject):
+        super(DisplayCavity, self).__init__(cavityNum, rackObject)
+        self.statusPV = PV(self.pvPrefix + STATUS_SUFFIX)
+        self.severityPV = PV(self.pvPrefix + SEVERITY_SUFFIX)
+
+    def runThroughFaults(self):
+        while True:
+            for fault in faults:
+                try:                               
+                    if fault.isFaulted(self):
+                        self.statusPV.put(fault.tlc)
+                        self.severityPV.put(fault.severity)
+                        break
+                    self.statusPV.put(self.number)
+                    self.severityPV.put(0)
+                except PvInvalid:
+                    self.statusPV.put("INV")
+                    self.severityPV.put(3)     
+
 DISPLAY_LINACS = []
 for name, cryomoduleList in LINACS.items():
     #print(name, cryomoduleList)
-    DISPLAY_LINACS.append(Linac(name, cryomoduleList, Cavity))
+    DISPLAY_LINACS.append(Linac(name, cryomoduleList, DisplayCavity))
+    
+
